refactor(domain): lift UnifiedDataRepository port into the domain layer (swamp-club#229)
Move the UnifiedDataRepository interface, OwnershipValidationError class, and
GarbageCollectionResult type from src/infrastructure/persistence/ into a new
src/domain/data/repositories.ts so the domain layer owns the contract. The
infra file imports them back and re-exports for non-domain consumers (cli,
libswamp, integration tests). FileSystemUnifiedDataRepository keeps its
implements clause against the now-domain-side interface.

Switch all 13 production domain importers (and 10 domain test files) of
UnifiedDataRepository to import from the new domain path. Update the two data
services (rename, delete) to also depend on the domain-side DefinitionRepository
instead of the concrete YamlDefinitionRepository — findDefinitionByIdOrName
already accepts the interface, so no other call-site changes are needed.

Decrement KNOWN_DOMAIN_INFRA_VIOLATIONS in integration/ddd_layer_rules_test.ts
from 31 to 22 (down 9 — files where UnifiedDataRepository was the sole
infrastructure import). Files that retain other infra imports (CatalogStore,
CatalogRow, YamlOutputRepository, FileSystemUnifiedDataRepository concrete
class) update their UnifiedDataRepository import path for canonical-source
correctness but stay on the violations list — those are deeper refactors out
of scope here.

Add src/infrastructure/persistence/unified_data_repository_re_export_test.ts
to lock the value re-export contract: asserts class identity is preserved
across both import paths and that instanceof works against either side. Guards
against a future regression that converts the value re-export to a type-only
re-export or duplicates the class definition.
