+++ src/domain/data/data_rename_service.ts
@@ -18,8 +18,8 @@
 // along with Swamp.  If not, see <https://www.gnu.org/licenses/>.
 
 import { DataMetadataSchema } from "./data_metadata.ts";
-import type { UnifiedDataRepository } from "../../infrastructure/persistence/unified_data_repository.ts";
-import type { YamlDefinitionRepository } from "../../infrastructure/persistence/yaml_definition_repository.ts";
+import type { UnifiedDataRepository } from "./repositories.ts";
+import type { DefinitionRepository } from "../definitions/repositories.ts";
 import { findDefinitionByIdOrName } from "../models/model_lookup.ts";
 
 /**
@@ -44,7 +44,7 @@ export interface RenameResult {
 export class DataRenameService {
   constructor(
     private readonly dataRepo: UnifiedDataRepository,
-    private readonly definitionRepo: YamlDefinitionRepository,
+    private readonly definitionRepo: DefinitionRepository,
   ) {}
 
   /**
