+++ src/domain/models/command/shell/shell_model_test.ts
@@ -28,7 +28,7 @@ import {
   shellModel,
 } from "./shell_model.ts";
 import type { DataHandle, DataWriter, MethodContext } from "../../model.ts";
-import type { UnifiedDataRepository } from "../../../../infrastructure/persistence/unified_data_repository.ts";
+import type { UnifiedDataRepository } from "../../../data/repositories.ts";
 import type { DefinitionRepository } from "../../../definitions/repositories.ts";
 import { type DataId, generateDataId } from "../../../data/data_id.ts";
 import { getLogger } from "@logtape/logtape";
