+++ src/domain/models/method_context_test.ts
@@ -31,7 +31,7 @@ import type { DefinitionRepository } from "../definitions/repositories.ts";
 import type { OutputRepository } from "./repositories.ts";
 import type { VaultService } from "../vaults/vault_service.ts";
 import type { SecretRedactor } from "../secrets/mod.ts";
-import type { UnifiedDataRepository } from "../../infrastructure/persistence/unified_data_repository.ts";
+import type { UnifiedDataRepository } from "../data/repositories.ts";
 
 function makeCommon(
   overrides: Partial<CommonMethodContextDeps> = {},
