+++ src/domain/models/method_execution_service_test.ts
@@ -29,7 +29,7 @@ import type {
   ModelDefinition,
 } from "./model.ts";
 import { z } from "zod";
-import type { UnifiedDataRepository } from "../../infrastructure/persistence/unified_data_repository.ts";
+import type { UnifiedDataRepository } from "../data/repositories.ts";
 import type { DefinitionRepository } from "../definitions/repositories.ts";
 import { type DataId, generateDataId } from "../data/data_id.ts";
 import { Data } from "../data/data.ts";
