+++ src/domain/models/model_test.ts
@@ -32,7 +32,7 @@ import {
 } from "./model.ts";
 import { ModelType } from "./model_type.ts";
 import { createDefinitionId } from "../definitions/definition.ts";
-import type { UnifiedDataRepository } from "../../infrastructure/persistence/unified_data_repository.ts";
+import type { UnifiedDataRepository } from "../data/repositories.ts";
 import type { DefinitionRepository } from "../definitions/repositories.ts";
 import { type DataId, generateDataId } from "../data/data_id.ts";
 import { getLogger } from "@logtape/logtape";
