+++ src/infrastructure/persistence/unified_data_repository.ts
@@ -35,389 +35,21 @@ import {
 import { ModelType } from "../../domain/models/model_type.ts";
 import type { MarkDirtyHook } from "../../domain/datastore/datastore_sync_service.ts";
 import type { CatalogStore } from "./catalog_store.ts";
+import {
+  type GarbageCollectionResult,
+  OwnershipValidationError,
+  type UnifiedDataRepository,
+} from "../../domain/data/repositories.ts";
+
+// Re-export domain repository types so existing infra-path importers keep working.
+// New domain code should import directly from src/domain/data/repositories.ts.
+export {
+  type GarbageCollectionResult,
+  OwnershipValidationError,
+  type UnifiedDataRepository,
+};
 
 const logger = getSwampLogger(["data", "repository"]);
-
-/**
- * Error thrown when ownership validation fails.
- */
-export class OwnershipValidationError extends Error {
-  constructor(
-    readonly dataName: string,
-    readonly existingOwner: { ownerType: string; ownerRef: string },
-    readonly newOwner: { ownerType: string; ownerRef: string },
-  ) {
-    super(
-      `Ownership validation failed for "${dataName}": ` +
-        `existing owner "${existingOwner.ownerType}:${existingOwner.ownerRef}" ` +
-        `does not match new owner "${newOwner.ownerType}:${newOwner.ownerRef}"`,
-    );
-    this.name = "OwnershipValidationError";
-  }
-}
-
-/**
- * Result of garbage collection operation.
- */
-export interface GarbageCollectionResult {
-  versionsRemoved: number;
-  bytesReclaimed: number;
-}
-
-/**
- * Repository interface for unified Data storage with versioning.
- */
-export interface UnifiedDataRepository {
-  /**
-   * Finds all data across all model types and models.
-   *
-   * @returns Array of data with their model type and model ID
-   */
-  findAllGlobal(): Promise<
-    Array<{ data: Data; modelType: ModelType; modelId: string }>
-  >;
-
-  /**
-   * Finds data by name, optionally for a specific version.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - Optional version (defaults to latest)
-   * @returns The data if found, or null
-   */
-  findByName(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version?: number,
-  ): Promise<Data | null>;
-
-  /**
-   * Finds data by ID, optionally for a specific version.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataId - The data ID
-   * @param version - Optional version (defaults to latest)
-   * @returns The data if found, or null
-   */
-  findById(
-    type: ModelType,
-    modelId: string,
-    dataId: DataId,
-    version?: number,
-  ): Promise<Data | null>;
-
-  /**
-   * Lists all versions for a data name.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @returns Array of version numbers in ascending order
-   */
-  listVersions(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-  ): Promise<number[]>;
-
-  /**
-   * Finds all data for a model.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @returns Array of data (latest version of each)
-   */
-  findAllForModel(type: ModelType, modelId: string): Promise<Data[]>;
-
-  /**
-   * Saves data with its content, creating a new version.
-   * Validates ownership if data with the same name already exists.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param data - The data entity
-   * @param content - The content to save
-   * @returns The saved version number
-   * @throws OwnershipValidationError if ownership validation fails
-   */
-  save(
-    type: ModelType,
-    modelId: string,
-    data: Data,
-    content: Uint8Array,
-  ): Promise<{ version: number }>;
-
-  /**
-   * Appends content to streaming data.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param content - The content to append
-   */
-  append(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    content: Uint8Array,
-  ): Promise<void>;
-
-  /**
-   * Streams content from data.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - Optional version (defaults to latest)
-   * @returns Async iterable of content chunks
-   */
-  stream(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version?: number,
-  ): AsyncIterable<Uint8Array>;
-
-  /**
-   * Gets the full content of data.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - Optional version (defaults to latest)
-   * @returns The content or null if not found
-   */
-  getContent(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version?: number,
-  ): Promise<Uint8Array | null>;
-
-  /**
-   * Deletes data, optionally for a specific version.
-   * If no version is specified, deletes all versions.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - Optional version to delete (all versions if not specified)
-   */
-  delete(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version?: number,
-  ): Promise<void>;
-
-  /**
-   * Removes the latest symlink for expired data (soft delete).
-   * Version directories remain on disk but data becomes inaccessible.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   */
-  removeLatestMarker(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-  ): Promise<void>;
-
-  /**
-   * Allocates a new version directory without writing content.
-   * Used by DataWriter for direct file I/O.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param data - The data entity (for ownership validation)
-   * @returns The allocated version number and content file path
-   */
-  allocateVersion(
-    type: ModelType,
-    modelId: string,
-    data: Data,
-  ): Promise<{ version: number; contentPath: string }>;
-
-  /**
-   * Finalizes a previously allocated version by writing metadata and updating symlinks.
-   * Content must already exist on disk at the content path.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param data - The data entity
-   * @param version - The version number to finalize
-   * @returns Size and checksum of the content
-   */
-  finalizeVersion(
-    type: ModelType,
-    modelId: string,
-    data: Data,
-    version: number,
-  ): Promise<{ size: number; checksum: string }>;
-
-  /**
-   * Generates a new unique ID.
-   */
-  nextId(): DataId;
-
-  /**
-   * Returns the directory path for a data version.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - The version number
-   * @returns The directory path
-   */
-  getPath(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version: number,
-  ): string;
-
-  /**
-   * Returns the content file path for a data version.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - The version number
-   * @returns The content file path
-   */
-  getContentPath(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version: number,
-  ): string;
-
-  /**
-   * Collects garbage according to each data's garbage collection policy.
-   *
-   * When `options.dryRun` is true, no versions are deleted — the returned
-   * `versionsRemoved` and `bytesReclaimed` reflect what would be removed.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param options - Options for the operation
-   * @returns The result of garbage collection
-   */
-  collectGarbage(
-    type: ModelType,
-    modelId: string,
-    options?: { dryRun?: boolean },
-  ): Promise<GarbageCollectionResult>;
-
-  /**
-   * Renames a data instance by copying the latest version to a new name
-   * and writing a tombstone with a forward reference on the old name.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param oldName - The current data name
-   * @param newName - The new data name
-   * @returns The rename result with version info
-   */
-  rename(
-    type: ModelType,
-    modelId: string,
-    oldName: string,
-    newName: string,
-  ): Promise<
-    {
-      oldName: string;
-      newName: string;
-      copiedVersion: number;
-      newVersion: number;
-    }
-  >;
-
-  // --- Sync read methods (for CEL expression evaluation) ---
-
-  /**
-   * Gets the latest version number synchronously.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @returns The latest version number, or null if not found
-   */
-  getLatestVersionSync(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-  ): number | null;
-
-  /**
-   * Finds data by name synchronously, optionally for a specific version.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - Optional version (defaults to latest)
-   * @returns The data if found, or null
-   */
-  findByNameSync(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version?: number,
-  ): Data | null;
-
-  /**
-   * Lists all versions for a data name synchronously.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @returns Array of version numbers in ascending order
-   */
-  listVersionsSync(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-  ): number[];
-
-  /**
-   * Gets the full content of data synchronously.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @param dataName - The data name
-   * @param version - Optional version (defaults to latest)
-   * @returns The content or null if not found
-   */
-  getContentSync(
-    type: ModelType,
-    modelId: string,
-    dataName: string,
-    version?: number,
-  ): Uint8Array | null;
-
-  /**
-   * Finds all data for a model synchronously.
-   *
-   * @param type - The model type
-   * @param modelId - The model input ID
-   * @returns Array of data (latest version of each)
-   */
-  findAllForModelSync(type: ModelType, modelId: string): Data[];
-
-  /**
-   * Finds all data across all model types and models synchronously.
-   * Used by DataQueryService for catalog backfill in sync contexts.
-   *
-   * @returns Array of data with their model type and model ID
-   */
-  findAllGlobalSync(): Array<
-    { data: Data; modelType: ModelType; modelId: string }
-  >;
-}
 
 /**
  * File system implementation of UnifiedDataRepository.
